Implemented SoundSetSource and FootstepSoundSource; fixed other few minor things
git-svn-id: http://sonetto.googlecode.com/svn@221 0bea0c47-0c44-0410-8422-6f9b8df7ced0

+++ branches/audiomanager/libsonetto/src/SonettoInputManager.cpp
@@ -64,7 +64,7 @@ namespace Sonetto
     // ----------------------------------------------------------------------
     void InputManager::initialize()
     {
-        for (size_t i = 1;i <= SDL_NumJoysticks();++i)
+        for (int i = 1;i <= SDL_NumJoysticks();++i)
         {
             mJoysticks.push_back(JoystickPtr(new Joystick(i)));
         }
